% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rims.R
\name{ct_luckyguy}
\alias{ct_luckyguy}
\title{Lucky guy function}
\usage{
ct_luckyguy(user_ans, etalon, algorithm = 1, ...)
}
\arguments{
\item{user_ans}{the vector of user responces}

\item{etalon}{the vector of etalon cathegories}

\item{algorithm}{1 means that is max_dist is set to + infinity,
2 means that iterative algorithm is used
1 is better for finding one-to-one correspondance,
2 is better to find many-to-one correspondance}

\item{translit}{logical whether to do transliteration}
}
\value{
pretty correspondance table if you are a lucky guy
#
}
\description{
This function tries to create a final correspondance table
starting from a vector of user answers and a vector of etalons
}
\details{
The Levenstein distance is used. When to matches give the same distance
the first one is returned. Warning: for lucky guys only!
}
\examples{
ct_luckyguy(user_ans,etalon,translit = FALSE)
ct_luckyguy(user_ans,ct_start)
}

