% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rims.R
\name{ct_new_block}
\alias{ct_new_block}
\title{Create additional correspondance table from user responses and actual correspondance table}
\usage{
ct_new_block(z, ct, max_dist = 1)
}
\arguments{
\item{z}{the vector of user responces}

\item{ct}{actual correspondance table}

\item{max_dist}{maximum Levenstein distance}
}
\value{
additional lines for correspondance table
#
}
\description{
Create additional correspondance table from user responses and actual correspondance table
}
\details{
This function uses Levenstein distance to create additional entries for correspondance table
}
\examples{
ct_new_block(z,ct)
}

