% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rims.R
\name{ct_start}
\alias{ct_start}
\title{Create base correspondance table from vector of etalon cathegory names}
\usage{
ct_start(etal_cat, add_original = FALSE, translit = TRUE)
}
\arguments{
\item{add_original}{whether we include trivial correspondance etalon <-> etalon}

\item{translit}{logical whether we transliterate while standartising responses}

\item{x}{the vector of etalon cathegory names}
}
\value{
data.frame with basic correspondance table
}
\description{
Create base correspondance table from vector of etalon cathegory names
}
\details{
This function creates basic correspondance table from vector of etalon cathegory names
}
\examples{
ct_start(c("Iphone","Samsung","HTC"))
}

