% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rims.R
\name{ct_unmatched}
\alias{ct_unmatched}
\title{Find unmatched user responses given correspondance table}
\usage{
ct_unmatched(z, ct)
}
\arguments{
\item{z}{the vector of user responces}

\item{ct}{actual correspondance table with in_cat and out_cat variables}
}
\value{
vector of unmatched user responses
}
\description{
Find unmatched user responses given correspondance table
}
\details{
This function finds unmatched user responses given correspondance table
}
\examples{
ct_unmatched(z,ct)
}

