% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rims.R
\name{one_factor}
\alias{one_factor}
\title{Replace two or more factor levels by the first one.}
\usage{
one_factor(x, levels)
}
\arguments{
\item{x}{the factor variable with misprints in level names}

\item{levels}{the levels that should be replaced by one level, levels[1] is used as the replacement.}
}
\value{
correctly encoded x
}
\description{
Replace two or more factor levels by the first one.
}
\details{
In raw data the same level of a factor variable may be encoded with errors, i.e. misprints.
In this case many levels may correspond actually to one true level.
This function replaces all mentioned levels by the first one.
}
\examples{
x <- c("Male","Female","male","Mle","Female")
unite_factors(x,c("Male","male","Mle"))
}

