% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rims.R
\docType{package}
\name{rims}
\alias{rims}
\alias{rims-package}
\title{small rims package}
\description{
This package is designed for easy life of IMS Health company.
}
\details{
To see the list of available functions click on index below!
Be happy, don't worry :)
}
\author{
Boris Demeshev
}

