% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rims.R
\name{str_stand}
\alias{str_stand}
\title{Standartize character vector}
\usage{
str_stand(z, translit = TRUE)
}
\arguments{
\item{translit}{logical whether we transliterate vector into latin}

\item{x}{the character vector}
}
\value{
standartized character vector
}
\description{
Standartize character vector
}
\details{
Standartize means: trimming white space at the end and in the beginning,
elimination of punctiation, transliteration, removing capitalization
}
\examples{
str_stand("пРивет!")
}

