% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rims.R
\name{str_translit}
\alias{str_translit}
\title{Transliterate cyrillic text}
\usage{
str_translit(x)
}
\arguments{
\item{x}{the vector of cyrillic characters}
}
\value{
transliterated vector
#
}
\description{
Transliterate cyrillic text
}
\details{
This function uses the transliteration tradition where "й" goes to "y"
}
\examples{
str_translit("привет")
}

