% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rims.R
\name{str_utf2cp}
\alias{str_utf2cp}
\title{Change encoding of cyrillic text from "utf8" to "cp1251"}
\usage{
str_utf2cp(x, backward = FALSE)
}
\arguments{
\item{x}{the vector or data.frame of cyrillic characters}

\item{backward}{(default FALSE) TRUE: utf8 to cp1251, FALSE: cp1251 to utf8}
}
\value{
reencoded vector or data.frame
}
\description{
Change encoding of cyrillic text from "utf8" to "cp1251"
}
\details{
This function changes encoding of a vector or data.frame from "utf8" to "cp1251"
}
\examples{
str_utf2cp("привет")
}

