% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rims.R
\name{str_utf2cp0}
\alias{str_utf2cp0}
\title{Change encoding of cyrillic text from "utf8" to "cp1251"}
\usage{
str_utf2cp0(x, backward = FALSE)
}
\arguments{
\item{x}{the vector of cyrillic characters}

\item{backward}{(default FALSE) TRUE: utf8 to cp1251, FALSE: cp1251 to utf8}
}
\value{
reencoded vector
}
\description{
Change encoding of cyrillic text from "utf8" to "cp1251"
}
\details{
This function changes encoding of a vector from "utf8" to "cp1251"
This function applies changes only to character and factor variables
}
\examples{
str_utf2cp0("привет")
}

