% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rims.R
\name{str_word}
\alias{str_word}
\title{Get specific word from a vector of character sentences}
\usage{
str_word(x, num = 1, remove.duplicate = TRUE)
}
\arguments{
\item{x}{the vector of cyrillic characters}

\item{num}{the number (numbers) of word, negative numbers mean "from the end of sentence"}

\item{remove.duplicate}{logical, indicates whether the same word should be reported once or twice}
}
\value{
data frame of words with numbers
#
}
\description{
Get specific word from a vector of character sentences
}
\details{
This function uses the transliteration tradition where "й" goes to "y"
}
\examples{
str_word(c("привет","Маша, это я, Дубровский"))
}

