% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rims.R
\name{write_rus_csv}
\alias{write_rus_csv}
\title{Write .csv file using Russian Excel standard.}
\usage{
write_rus_csv(obj, file)
}
\arguments{
\item{obj}{data.frame to be written}

\item{file}{file name}
}
\description{
Write .csv file using Russian Excel standard.
}
\details{
Write .csv file using Russian Excel standard.
}
\examples{
write_rus_csv(cars, file="cars.csv")
}

